[fs] overlayfs: Warn on copy up if a process has a R/O fd open to the lower file
Message-id: <30119.1432909751@warthog.procyon.org.uk>
Patchwork-id: 114607
O-Subject: [RHEL7.2 PATCH] overlayfs: Warn on copy up if a process has a R/O fd open to the lower file [BZ 1226346]
Bugzilla: 1226346
RH-Acked-by: Jeremy Eder <jeder@redhat.com>
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Print a warning when overlayfs copies up a file if the process that triggered
the copy up has a R/O fd open to the lower file being copied up.

This can help catch applications that do things like the following:

	fd1 = open("foo", O_RDONLY);
	fd2 = open("foo", O_RDWR);

where they expect fd1 and fd2 to refer to the same file - which will no longer
be the case post-copy up.

With this patch, the following commands:

	bash 5</mnt/a/foo128
	6<>/mnt/a/foo128

assuming /mnt/a/foo128 to be an un-copied up file on an overlay will produce
the following warning in the kernel log:

	overlayfs: Copying up foo128, but open R/O on fd 5

This is enabled by setting:

	/sys/module/overlay/parameters/check_copy_up

to 1.

Signed-off-by: David Howells <dhowells@redhat.com>
RH-Bugzilla: 1226346
Signed-off-by: Rafael Aquini <aquini@redhat.com>