[fs] gfs2: handle NULL rgd in set_rgrp_preferences
Message-id: <1433365454-64416-1-git-send-email-adas@redhat.com>
Patchwork-id: 115394
O-Subject: [RHEL7.2 PATCH] gfs2: handle NULL rgd in set_rgrp_preferences
Bugzilla: 1211663
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The function set_rgrp_preferences() does not handle the (rarely
returned) NULL value from gfs2_rgrpd_get_next() and this patch
fixes that.

The fs image in question is only 150MB in size which allows for
only 1 rgrp to be created. The in-memory rb tree has only 1 node
and when gfs2_rgrpd_get_next() is called on this sole rgrp, it
returns NULL. (Default behavior is to wrap around the rb tree and
return the first node to give the illusion of a circular linked
list. In the case of only 1 rgrp, we can't have
gfs2_rgrpd_get_next() return the same rgrp (first, last, next all
point to the same rgrp)... that would cause unintended consequences
and infinite loops.)

This patch is already upstream:
http://git.kernel.org/cgit/linux/kernel/git/gfs2/linux-gfs2.git/commit/?h=for-next&id=959b6717175713259664950f3bba2418b038f69a

Resolves: rhbz#1211663
Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>