[kvm] x86: pvclock: Really remove the sched notifier for cross-cpu migrations
Message-id: <1433759435-30449-13-git-send-email-pbonzini@redhat.com>
Patchwork-id: 115654
O-Subject: [RHEL7.2 PATCH 12/70] x86: pvclock: Really remove the sched notifier for cross-cpu migrations
Bugzilla: 1202825
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Radim Krcmar <rkrcmar@redhat.com>
RH-Acked-by: Bandan Das <bsd@redhat.com>
RH-Acked-by: Igor Mammedov <imammedo@redhat.com>

This reverts commits 0a4e6be9ca17c54817cf814b4b5aa60478c6df27
and 80f7fdb1c7f0f9266421f823964fd1962681f6ce.

The task migration notifier was originally introduced in order to support
the pvclock vsyscall with non-synchronized TSC, but KVM only supports it
with synchronized TSC.  Hence, on KVM the race condition is only needed
due to a bad implementation on the host side, and even then it's so rare
that it's mostly theoretical.

As far as KVM is concerned it's possible to fix the host, avoiding the
additional complexity in the vDSO and the (re)introduction of the task
migration notifier.

Xen, on the other hand, hasn't yet implemented vsyscall support at
all, so we do not care about its plans for non-synchronized TSC.

Reported-by: Peter Zijlstra <peterz@infradead.org>
Suggested-by: Marcelo Tosatti <mtosatti@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
(cherry picked from commit 73459e2a1ada09a68c02cc5b73f3116fc8194b3d)
Signed-off-by: Rafael Aquini <aquini@redhat.com>