[fs] gfs2: fix quota updates on block boundaries
Message-id: <1433777958-48638-1-git-send-email-adas@redhat.com>
Patchwork-id: 115901
O-Subject: [RHEL7.2 v3 PATCH 5/6] gfs2: fix quota updates on block boundaries
Bugzilla: 1174295
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

For smaller block sizes (512B, 1K, 2K), some quotas straddle block
boundaries such that the usage value is on one block and the rest
of the quota is on the previous block. In such cases, the value
does not get updated correctly. This patch fixes that by addressing
the boundary conditions correctly.

This patch also adds a (s64) cast that was missing in a call to
gfs2_quota_change() in inode.c. Another (s64) cast was missing in quota.c.
This version adds that.

Resolves: rhbz#1174295
Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>