[alsa] ctxfi: added reference of snd_card
Message-id: <143386227049-50-patchset-jkysela@redhat.com>
Patchwork-id: 116261
O-Subject: [RHEL 7.2 PATCH 049/763] ALSA: ctxfi: added reference of snd_card
Bugzilla: 1197064
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

added a pointer of snd_card in some of the structures to get a
reference of the card from other functions.
these references of snd_card will be initialised in the next patch
of this series and as of now these snd_card will be used to print the
the device information when we convert the pr_* macros to dev_* in a
later patch of this series.

Signed-off-by: Sudip Mukherjee <sudip@vectorindia.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 3d0fdc86e4b500dfcfbf2f68039d2d6853536c2e)
rhbz#1197064
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>