[alsa] pcm: Fix referred substream in snd_pcm_action_group() unlock loop
Message-id: <143386227049-61-patchset-jkysela@redhat.com>
Patchwork-id: 116246
O-Subject: [RHEL 7.2 PATCH 060/763] ALSA: pcm: Fix referred substream in snd_pcm_action_group() unlock loop
Bugzilla: 1197064
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

In the unlock loop of snd_pcm_action_group(), the object "s" is used
as the check of nonatomic PCM, but it should be rather "s1", which is
the iterator of the loop.  This supposedly causes a kernel panic when
the substreams in operatino are linked.

Fixes: 257f8cce5d40 ('ALSA: pcm: Allow nonatomic trigger operations')
Reported-and-tested-by: Arthur Marsh <arthur.marsh@internode.on.net>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 811deedebab38f8360a700a52b0b75688c9a10f7)
rhbz#1197064
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>