[alsa] au88x0: pr_* replaced with dev_*
Message-id: <143386227049-64-patchset-jkysela@redhat.com>
Patchwork-id: 116262
O-Subject: [RHEL 7.2 PATCH 063/763] ALSA: au88x0: pr_* replaced with dev_*
Bugzilla: 1197064
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

pr_* macros replaced with dev_* as they are more preffered over pr_*.
each file which had pr_* was reviewed manually and replaced with dev_*.
here we have actually used the reference of the vortex which was added
to some functions in the previous patch of this series.

The prefix of the CARD_NAME and prefix of "vortex:" was also
removed as the dev_* will print the device name.

Signed-off-by: Sudip Mukherjee <sudip@vectorindia.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 70c84418bf74f582e29906f1eeb19f2e9da53ddd)
rhbz#1197064
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>