[alsa] hda - Add workaround for CMI8888 snoop behavior
Message-id: <143386227049-127-patchset-jkysela@redhat.com>
Patchwork-id: 116286
O-Subject: [RHEL 7.2 PATCH 126/763] ALSA: hda - Add workaround for CMI8888 snoop behavior
Bugzilla: 1197064
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

CMI8888 shows the stuttering playback when the snooping is disabled
on the audio buffer.  Meanwhile, we've got reports that CORB/RIRB
doesn't work in the snooped mode.  So, as a compromise, disable the
snoop only for CORB/RIRB and enable the snoop for the stream buffers.

The resultant patch became a bit ugly, unfortunately, but we still can
live with it.

Reported-and-tested-by: Geoffrey McRae <geoff@spacevs.com>
Cc: <stable@vger.kernel.org> # 3.17+
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(picked from kernel: 3b70bdba2fcb374a2235a56ab73334348d819579)
rhbz#1197064
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>