[alsa] regmap: rbtree: Simplify adjacent node look-up
Message-id: <143386227049-757-patchset-jkysela@redhat.com>
Patchwork-id: 116913
O-Subject: [RHEL 7.2 PATCH 756/763] ALSA: regmap: rbtree: Simplify adjacent node look-up
Bugzilla: 1197064
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

A register which is adjacent to a node will either be left to the first
register or right to the last register. It will not be within the node's range,
so there is no point in checking for each register cached by the node whether
the new register is next to it. It is sufficient to check whether the register
comes before the first register or after the last register of the node.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Mark Brown <broonie@linaro.org>

(picked from kernel: 194c753a214ba7f1497552dd530021884d164146)
rhbz#1197064
Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>