[char] ipmi: Remove the now unnecessary message queue
Message-id: <1433875773-31976-14-git-send-email-tcamuso@redhat.com>
Patchwork-id: 117039
O-Subject: [PATCH RHEL7.2 13/43] ipmi: Remove the now unnecessary message queue
Bugzilla: 1229675
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1229675

Cherry-picked from the following upstream commit.

commit b874b985c816c74a9bda04082f18db88dcbc808a
Author: Corey Minyard <cminyard@mvista.com>
Date:   Thu Nov 6 17:01:59 2014 -0600

    ipmi: Remove the now unnecessary message queue

    A message queue was added to the message handler, so the SMI
    interfaces only need to handle one message at a time.  Pull out
    the message queue.  This also leads to some significant
    simplification in the shutdown of an interface, since the
    message handler now does a lot of the cleanup.

    Signed-off-by: Corey Minyard <cminyard@mvista.com>

Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
smi_info.interrupt_disabled was changed to bool in an earlier patch.
Assigning it to a bool value, instead of to an int, caused it to be
flagged as a conflict. Upstream was not consistent to change the
assignment of this field to boolean values elsewhere in the code.
Perhaps in a future upstream commit.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>