[md] raid5: need_this_block: tidy/fix last condition
Message-id: <1433875910-23473-36-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 117135
O-Subject: [RHEL7.2 PATCH 035/100] md/raid5: need_this_block: tidy/fix last condition.
Bugzilla: 1150149 1173510 1194720
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.de>

That last condition is unclear and over cautious.

There are two related issues here.

If a partial write is destined for a missing device, then
either RMW or RCW can work.  We must read all the available
block.  Only then can the missing blocks be calculated, and
then the parity update performed.

If RMW is not an option, then there is a complication even
without partial writes.  If we would need to read a missing
device to perform the reconstruction, then we must first read every
block so the missing device data can be computed.
This is the case for RAID6 (Which currently does not support
RMW) and for times when we don't trust the parity (after a crash)
and so are in the process of resyncing it.

So make these two cases more clear and separate, and perform
the relevant tests more  thoroughly.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit ea664c8245f3d5e78d05d1250bc0be0d60e264af)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>