[md] raid5: Ensure a batch member is not handled prematurely
Message-id: <1433875910-23473-94-git-send-email-Jes.Sorensen@redhat.com>
Patchwork-id: 117125
O-Subject: [RHEL7.2 PATCH 093/100] md/raid5: Ensure a batch member is not handled prematurely.
Bugzilla: 1150149 1173510 1194720
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: NeilBrown <neilb@suse.de>

If a stripe is a member of a batch, but not the head, it must
not be handled separately from the rest of the batch.

'clear_batch_ready()' handles this requirement to some
extent but not completely.  If a member is passed to handle_stripe()
a second time it returns '0' indicating the stripe can be handled,
which is wrong.
So add an extra test.

Signed-off-by: NeilBrown <neilb@suse.de>
(cherry picked from commit b15a9dbdbfe72848b7ed4cd3f97fe80daaf99c89)
Signed-off-by: Jes Sorensen <Jes.Sorensen@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>