[infiniband] ib/core: Fix XRC race condition in ib_uverbs_open_qp
Message-id: <9510d3200a8c599fb137a823cc9e31f737ca4d97.1433950569.git.honli@redhat.com>
Patchwork-id: 117458
O-Subject: [RHEL-7.2 PATCH 04/34] IB/core: Fix XRC race condition in ib_uverbs_open_qp
Bugzilla: 1229265
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Doug Ledford <dledford@redhat.com>

From: Jack Morgenstein <jackm@dev.mellanox.co.il>

In ib_uverbs_open_qp, the sharable xrc target qp is created as a
"pseudo" qp and added to a list of qp's sharing the same physical
QP.  This is done before the "pseudo" qp is assigned a uobject.

There is a race condition here if an async event arrives at the
physical qp.  If the event is handled after the pseudo qp is added to
the list, but before it is assigned a uobject, the kernel crashes in
ib_uverbs_qp_event_handler, due to trying to dereference a NULL
uobject pointer.

Note that simply checking for non-NULL is not enough, due to error
flows in ib_uverbs_open_qp.  If the failure is after assigning the
uobject, but before the qp has fully been created, we still have a
problem.

Thus, in ib_uverbs_qp_event_handler, we test that the uobject is
present, and also that it is live.

Reported-by: Matthew Finlay <matt@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: Roland Dreier <roland@purestorage.com>
(cherry picked from commit a040f95dc81986d7d55bd98a9455217522ef706d)
Signed-off-by: Honggang Li <honli@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>